Move roundForDisplay to separate module to reduce bundle size
Move roundForDisplay() function from packages/utils/src/math/math.ts to a new
packages/utils/src/math/rounding.ts module. This keeps the math-expressions
dependency isolated and prevents consumers importing only mathjaxConfig from
unnecessarily pulling the heavy math-expressions runtime into their bundles.

API compatibility preserved: roundForDisplay remains re-exported from @doenet/utils
so existing imports continue to work.
