+++ packages/utils/src/index.ts
@@ -12,6 +12,7 @@ export * from "./components/size";
 export * from "./math/array";
 export * from "./math/math";
 export * from "./math/mathexpressions";
+export * from "./math/rounding";
 export * from "./math/subset-of-reals-operations";
 export * from "./style/colorWords";
 export * from "./style/style";
