+++ packages/utils/src/math/math.ts
@@ -1,6 +1,3 @@
-import me from "math-expressions";
-import type { Expression, Tree } from "math-expressions";
-
 export const mathjaxConfig = {
     tex: {
         tags: "ams",
@@ -19,35 +16,3 @@ export const mathjaxConfig = {
         displayOverflow: "linebreak",
     },
 };
-
-type RoundingDependencyValues = {
-    displayDigits: number;
-    displayDecimals: number;
-    displaySmallAsZero: number;
-};
-
-export function roundForDisplay({
-    value,
-    dependencyValues,
-}: {
-    value: Expression | Tree;
-    dependencyValues: RoundingDependencyValues;
-}): Expression {
-    let rounded = me.round_numbers_to_precision_plus_decimals(
-        value,
-        dependencyValues.displayDigits,
-        dependencyValues.displayDecimals,
-    );
-
-    if (
-        dependencyValues.displayDigits > 0 &&
-        dependencyValues.displaySmallAsZero > 0
-    ) {
-        rounded = me.set_small_zero(
-            rounded,
-            dependencyValues.displaySmallAsZero,
-        );
-    }
-
-    return rounded;
-}
